Pre-Test SmrtCoin by AleGio

+++ contracts/smrtCoin.sol
@@ -12,7 +12,7 @@ pragma solidity ^0.6.12;
  contract smrtCoin {
 
     //variables
-    mapping (address => uint) balances;
+    mapping (address => uint256) balances;
 
     //constructor
     constructor() public{
@@ -21,14 +21,14 @@ pragma solidity ^0.6.12;
 
 
     //Metodos
-    function send(address receiver, uint amount) public {
+    function send(address receiver, uint256 amount) public {
         require(balances[msg.sender] > amount);
         balances[msg.sender] -= amount;
         balances[receiver] += amount;
     }
 
-    function getBalance(address adr) public view returns(uint){
+    function getBalance(address adr) public view returns(uint256){
         return balances[adr];
     }
  }
-// 1:23:29
+// 1:23:29 --> 1:42:18
